Factor out SHashTable to utils/hash_table
