+++ tests/test_utils.c
@@ -20,86 +20,13 @@
 #define _GNU_SOURCE
 #define _FILE_OFFSET_BITS 64
 
-#include "dvb.h"
-#include "minisatip.h"
-#include "socketworks.h"
 #include "utils.h"
 #include "utils/testing.h"
-#include <arpa/inet.h>
-#include <ctype.h>
-#include <errno.h>
-#include <fcntl.h>
-#include <math.h>
-#include <net/if.h>
-#include <netdb.h>
-#include <netinet/in.h>
-#include <signal.h>
-#include <stdarg.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
+#include "opts.h"
 #include <string.h>
-#include <sys/ioctl.h>
-#include <sys/mman.h>
-#include <sys/socket.h>
 #include <sys/stat.h>
-#include <sys/types.h>
-#include <sys/ucontext.h>
-#include <syslog.h>
-#include <time.h>
-#include <unistd.h>
 
-/*
-// external references of utils.c
-int run_loop;
-char pid_file[50];
-__thread pthread_t tid;
-char app_name[] = "test";
-__thread char *strcpy(thread_name, "test");;
-
-struct struct_opts opts;
-*/
-
-#define MAX_HASH 100
-#define KEY_FOR_INDEX(i) (i)
-
-int test_hash_table() {
-    SHashTable h;
-    int i;
-    ASSERT(create_hash_table(&h, MAX_HASH - 10) == 0,
-           "expected to create hash successfully");
-    for (i = 0; i < MAX_HASH; i++) {
-        ASSERT(0 == setItem(&h, KEY_FOR_INDEX(i), &i, sizeof(i)),
-               "setItem failed");
-    }
-
-    ASSERT(0 == delItem(&h, KEY_FOR_INDEX(0)), "delItem failed");
-    ASSERT(NULL == getItem(&h, KEY_FOR_INDEX(0)),
-           "getItem is expected to fail after delete");
-    for (i = 1; i < MAX_HASH; i++) {
-        int *p = getItem(&h, KEY_FOR_INDEX(i));
-        ASSERT(p != NULL, "getItem should not fail");
-        ASSERT(i == *p, "getItem should return the correct key");
-        ASSERT(0 == delItem(&h, KEY_FOR_INDEX(i)),
-               "delItem should work for existing items");
-    }
-
-    for (i = 0; i < h.size; i++)
-        if (h.items[i].len > 0)
-            ASSERT(h.items[i].len, "len > 0 for existing items");
-
-    LOG("Hash conflicts %d, len %d, size %d", h.conflicts, h.len, h.size);
-
-    // no copy
-    _setItem(&h, 0, (void *)23, 4, 0);
-    i = 23;
-    ASSERT(23 == (uintptr_t)getItem(&h, 0), "expected the correct item");
-    setItem(&h, 0, &i, sizeof(i));
-
-    free_hash(&h);
-
-    return 0;
-}
+extern __thread char thread_name[100];
 
 int test_mkdir_recursive() {
     const char *dir = "/tmp/minisatip/test";
@@ -114,7 +41,6 @@ int test_mkdir_recursive() {
 int main() {
     opts.log = 255;
     strcpy(thread_name, "test");
-    TEST_FUNC(test_hash_table(), "testing hash table");
     TEST_FUNC(test_mkdir_recursive(), "testing directory creation");
     return 0;
 }
